Fix error when buildout.cfg is already there

+++ rt/commands/project.py
@@ -7,6 +7,7 @@ from fabric.api import run
 from fabric.context_managers import quiet
 from fabric.contrib import console
 from rt.commands.components import plone
+import os
 
 
 def supervisor(action='status'):
@@ -33,13 +34,14 @@ def component_buildout(components='*', profile='production.cfg'):
 
     for component in components:
         opts = {'component_dir': '%s/components/%s' % (local_buildout, component),
-                'profile' : profile}
+                'profile': profile}
         if not console.confirm("Do you want to launch buildout in %(component_dir)s" % opts):
             continue
 
         with lcd('%(component_dir)s' % opts):
             # symlink profile
-            local('ln -s ./profiles/%(profile)s buildout.cfg' % opts)
+            if not os.path.exists('buildout.cfg'):
+                local('ln -s ./profiles/%(profile)s buildout.cfg' % opts)
             # bootstrap
             local('../../bin/python bootstrap.py')
             # run buildout
@@ -58,4 +60,3 @@ def install_logrotate(components='*'):
     Install logrotate for components
     '''
     return plone.install_logrotate()
-
